--- a/sdk/metric/pipeline.go
+++ b/sdk/metric/pipeline.go
@@ -360,9 +360,8 @@ func (i *inserter[N]) cachedAggregator(scope instrumentation.Scope, kind Instrum
 	normID := id.normalize()
 	cv := i.aggregators.Lookup(normID, func() aggVal[N] {
 		b := aggregate.Builder[N]{
-			Temporality:               i.pipeline.reader.temporality(kind),
-			ExemplarReservoirProvider: stream.ExemplarReservoirProviderSelector(stream.Aggregation),
-			ExemplarFilter:            i.pipeline.exemplarFilter,
+			Temporality:   i.pipeline.reader.temporality(kind),
+			ReservoirFunc: reservoirFunc[N](stream.ExemplarReservoirProviderSelector(stream.Aggregation), i.pipeline.exemplarFilter),
 		}
 		b.Filter = stream.AttributeFilter
 		// A value less than or equal to zero will disable the aggregation