@dashpole 2024-10-23
It is probably from replacing nil with AlwaysOffFilter, if I had to guess.